@@ -1,6 +1,6 @@
 package de.intension.acronym;
 
-import javax.validation.constraints.NotNull;
+import jakarta.validation.constraints.NotNull;
 
 public final class AcronymUtil {
 
