@@ -22,13 +22,11 @@ import org.apache.http.message.BasicNameValuePair;
 import org.apache.http.util.EntityUtils;
 import org.jboss.logging.Logger;
 import org.keycloak.OAuth2Constants;
+import org.keycloak.representations.AccessTokenResponse;
 import org.keycloak.utils.MediaType;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-
-import twitter4j.JSONException;
-import twitter4j.JSONObject;
 
 public class IdPAssignmentsClient
     implements Closeable
@@ -59,7 +57,7 @@ public class IdPAssignmentsClient
      * Get list of allowed identity providers assigned to the given clientId.
      */
     public List<String> getListOfAllowedIdPs(String clientId, String apiRealm, String apiClientId, String apiClientSecret)
-        throws JSONException, IOException, URISyntaxException
+            throws IOException, URISyntaxException
     {
         List<String> listOfIdps;
         HttpGet httpGet = new HttpGet(new URIBuilder(String.format(restApiUrl, clientId)).build());
@@ -92,7 +90,7 @@ public class IdPAssignmentsClient
      * Calls the token endpoint to get a valid token that can be used to send the events to IMS.
      */
     private String getAccessToken(String realm, String clientId, String clientSecret)
-        throws JSONException, IOException
+            throws IOException
     {
         String tokenUrl = String.format("%s/realms/%s/protocol/openid-connect/token", kcAuthUrl, realm);
 
@@ -107,15 +105,11 @@ public class IdPAssignmentsClient
 
         httpPost.setEntity(new UrlEncodedFormEntity(params));
 
-        CloseableHttpResponse response = httpClient.execute(httpPost);
-
-        LOG.debugv("Get access token. Status = {0, number, integer}", response.getStatusLine().getStatusCode());
-
-        JSONObject readEntity = new JSONObject(EntityUtils.toString(response.getEntity()));
-
-        response.close();
-
-        return readEntity.getString(OAuth2Constants.ACCESS_TOKEN);
+        try (CloseableHttpResponse response = httpClient.execute(httpPost)) {
+            LOG.debugv("Get access token. Status = {0, number, integer}", response.getStatusLine().getStatusCode());
+            AccessTokenResponse accesstoken = objectMapper.readValue(EntityUtils.toString(response.getEntity()), AccessTokenResponse.class);
+            return accesstoken.getToken();
+        }
     }
 
     /**
